differentiate JSON Schema objects and types generated from schemas
