--- src/typescript-generator/Untitled-1
@@ -1,3 +1,0 @@
-
-
-
